Tests for tune_intensity
