--- tests/workup/intensity/2018-11-30/old.crv
@@ -1,76 +1,0 @@
-600						
-OPA/NOPA						
-0						
-0						
--1						
-0						
-0						
-4						
-0	1	2	3			
-2						
-NON-NON-NON-Sig						
-1						
-						
-1						
-795						
-4						
-0	0	0	0			
-25
-795.000000	1140.000000	4	-9.324437	2.165817	7.259192	56.780087	
-795.000000	1160.000000	4	-9.160146	2.187580	7.040502	56.874502	
-795.000000	1180.000000	4	-8.986959	2.208514	6.840363	56.963150	
-795.000000	1200.000000	4	-8.804875	2.228620	6.647657	57.046032	
-795.000000	1220.000000	4	-8.613894	2.247896	6.454240	57.123147	
-795.000000	1240.000000	4	-8.414017	2.266344	6.254239	57.194495	
-795.000000	1260.000000	4	-8.205244	2.283963	6.042106	57.260077	
-795.000000	1280.000000	4	-7.987574	2.300754	5.814651	57.319892	
-795.000000	1300.000000	4	-7.761008	2.316715	5.572952	57.373940	
-795.000000	1320.000000	4	-7.525545	2.331848	5.320444	57.422221	
-795.000000	1340.000000	4	-7.281186	2.346152	5.061374	57.464736	
-795.000000	1360.000000	4	-7.027931	2.359628	4.799532	57.501484	
-795.000000	1380.000000	4	-6.765779	2.372275	4.535686	57.532465	
-795.000000	1400.000000	4	-6.494730	2.384092	4.267147	57.557679	
-795.000000	1420.000000	4	-6.214785	2.395082	3.990444	57.577127	
-795.000000	1440.000000	4	-5.925944	2.405242	3.701662	57.590808	
-795.000000	1460.000000	4	-5.628206	2.414574	3.396541	57.598722	
-795.000000	1480.000000	4	-5.321572	2.423076	3.074929	57.600870	
-795.000000	1500.000000	4	-5.006041	2.430751	2.747495	57.597250	
-795.000000	1520.000000	4	-4.681614	2.437596	2.426308	57.587864	
-795.000000	1540.000000	4	-4.348291	2.443613	2.119391	57.572712	
-795.000000	1560.000000	4	-4.006070	2.448801	1.830076	57.551792	
-795.000000	1580.000000	4	-3.654954	2.453160	1.562761	57.525106	
-795.000000	1600.000000	4	-3.294941	2.456690	1.313887	57.492653	
-795.000000	1620.000000	4	-2.926032	2.459392	1.079490	57.454434	
-NON-NON-NON-Idl						
-1						
-						
-0						
-795						
-4						
-0	0	0	0			
-25
-795.000000	1561.090909	4	-2.926032	2.459392	1.079490	57.454434	
-795.000000	1580.124224	4	-3.294941	2.456690	1.313887	57.492653	
-795.000000	1600.127389	4	-3.654954	2.453160	1.562761	57.525106	
-795.000000	1621.176471	4	-4.006070	2.448801	1.830076	57.551792	
-795.000000	1643.355705	4	-4.348291	2.443613	2.119391	57.572712	
-795.000000	1666.758621	4	-4.681614	2.437596	2.426308	57.587864	
-795.000000	1691.489362	4	-5.006041	2.430751	2.747495	57.597250	
-795.000000	1717.664234	4	-5.321572	2.423076	3.074929	57.600870	
-795.000000	1745.413534	4	-5.628206	2.414574	3.396541	57.598722	
-795.000000	1774.883721	4	-5.925944	2.405242	3.701662	57.590808	
-795.000000	1806.240000	4	-6.214785	2.395082	3.990444	57.577127	
-795.000000	1839.669421	4	-6.494730	2.384092	4.267147	57.557679	
-795.000000	1875.384615	4	-6.765779	2.372275	4.535686	57.532465	
-795.000000	1913.628319	4	-7.027931	2.359628	4.799532	57.501484	
-795.000000	1954.678899	4	-7.281186	2.346152	5.061374	57.464736	
-795.000000	1998.857143	4	-7.525545	2.331848	5.320444	57.422221	
-795.000000	2046.534653	4	-7.761008	2.316715	5.572952	57.373940	
-795.000000	2098.144330	4	-7.987574	2.300754	5.814651	57.319892	
-795.000000	2154.193548	4	-8.205244	2.283963	6.042106	57.260077	
-795.000000	2215.280899	4	-8.414017	2.266344	6.254239	57.194495	
-795.000000	2282.117647	4	-8.613894	2.247896	6.454240	57.123147	
-795.000000	2355.555556	4	-8.804875	2.228620	6.647657	57.046032	
-795.000000	2436.623377	4	-8.986959	2.208514	6.840363	56.963150	
-795.000000	2526.575342	4	-9.160146	2.187580	7.040502	56.874502	
-795.000000	2626.956522	4	-9.324437	2.165817	7.259192	56.780087
