+++ tests/workup/intensity/gtol_ltol_instr.json
@@ -1,0 +1,1 @@
+{"name": null, "arrangements": {"sig": {"name": "sig", "tunes": {"c1": {"independent": [1140.0, 1160.0, 1180.0, 1200.0, 1220.0, 1240.0, 1260.0, 1280.0, 1300.0, 1320.0, 1340.0, 1360.0, 1380.0, 1400.0, 1420.0, 1440.0, 1460.0, 1480.0, 1500.0, 1520.0, 1540.0, 1560.0, 1580.0, 1600.0, 1620.0], "dependent": [-9.324437, -9.160146000000001, -8.986958999999999, -8.804875, -8.613894, -8.414017, -8.205244, -7.987574, -7.761008, -7.525545, -7.281186, -7.027931, -6.765778999999999, -6.494730000000001, -6.214785, -5.925944, -5.628206, -5.321572, -5.006041000000001, -4.681614, -4.348291000000001, -4.006069999999999, -3.654954, -3.2949410000000006, -2.926032], "ind_units": "nm", "dep_units": null}, "d1": {"independent": [1140.0, 1160.0, 1180.0, 1200.0, 1220.0, 1240.0, 1260.0, 1280.0, 1300.0, 1320.0, 1340.0, 1360.0, 1380.0, 1400.0, 1420.0, 1440.0, 1460.0, 1480.0, 1500.0, 1520.0, 1540.0, 1560.0, 1580.0, 1600.0, 1620.0], "dependent": [2.1658169999999997, 2.18758, 2.208514, 2.2286200000000003, 2.247896, 2.2663439999999997, 2.283963, 2.300754, 2.3167150000000003, 2.331848, 2.346152, 2.359628, 2.372275, 2.384092, 2.395082, 2.405242, 2.414574, 2.423076, 2.430751, 2.437596, 2.443613, 2.448801, 2.45316, 2.45669, 2.459392], "ind_units": "nm", "dep_units": null}, "c2": {"independent": [1140.0, 1160.0, 1180.0, 1200.0, 1220.0, 1240.0, 1260.0, 1280.0, 1300.0, 1320.0, 1340.0, 1360.0, 1380.0, 1400.0, 1420.0, 1440.0, 1460.0, 1480.0, 1500.0, 1520.0, 1540.0, 1560.0, 1580.0, 1600.0, 1620.0], "dependent": [7.259192, 7.040502, 6.840363000000001, 6.647657000000001, 6.45424, 6.254239, 6.0421059999999995, 5.8146510000000005, 5.572952, 5.320444, 5.061374, 4.799532, 4.535686, 4.2671470000000005, 3.990444, 3.7016620000000002, 3.3965410000000005, 3.074929, 2.7474950000000002, 2.426308, 2.119391, 1.8300759999999998, 1.562761, 1.313887, 1.07949], "ind_units": "nm", "dep_units": null}, "d2": {"independent": [1140.0, 1160.0, 1180.0, 1200.0, 1220.0, 1240.0, 1260.0, 1280.0, 1300.0, 1320.0, 1340.0, 1360.0, 1380.0, 1400.0, 1420.0, 1440.0, 1460.0, 1480.0, 1500.0, 1520.0, 1540.0, 1560.0, 1580.0, 1600.0, 1620.0], "dependent": [56.900087, 56.874502, 56.90315, 57.046032, 57.063146999999994, 57.194494999999996, 57.260077, 57.319892, 57.373940000000005, 57.482220999999996, 57.404736, 57.381484, 57.592465000000004, 57.557679, 57.697127, 57.650808, 57.658722, 57.840869999999995, 57.657250000000005, 57.647864000000006, 57.692712, 57.551792000000006, 56.20510600000001, 58.392653, 58.774434], "ind_units": "nm", "dep_units": null}}}}, "setables": {"c1": {"name": "c1"}, "d1": {"name": "d1"}, "c2": {"name": "c2"}, "d2": {"name": "d2"}}, "transition": {"type": "intensity", "metadata": {"channel": -1, "arrangement": "sig", "tune": "d2", "level": false, "gtol": 0.1, "ltol": 0.99999, "spline_kwargs": {"spline": false}}}}
