+++ tests/workup/intensity/test_intensity.py
@@ -1,0 +1,38 @@
+import attune
+import pathlib
+
+import numpy as np
+import WrightTools as wt
+
+
+__here__ = pathlib.Path(__file__).parent
+
+
+def test():
+    data = wt.open(__here__ / "data.wt5")
+    data.transform("w1=wm", "w1_Delay_2", "wa_points")
+    data.moment("wa_points", moment=0)
+    data.transform("w1=wm", "w1_Delay_2_points")
+    old = attune.open("input_instr.json")
+    new = attune.intensity(data, -1, "sig", "d2", old, autosave=False)
+    print(new)
+
+
+def test_ltol_with_gtol():
+    data = wt.open(__here__ / "data.wt5")
+    data.print_tree()
+    data.transform("w1=wm", "w1_Delay_2", "wa_points")
+    data.moment("wa_points", moment=0)
+    data.transform("w1=wm", "w1_Delay_2_points")
+    old = attune.open("input_instr.json")
+    new = attune.intensity(
+        data, -1, "sig", "d2", gtol=0.1, ltol=0.99999, spline=False, autosave=False, curve=old
+    )
+    correct = attune.open("gtol_ltol_instr.json")
+    assert np.allclose(new["sig"]["d2"].dependent, correct["sig"]["d2"].dependent, atol=0.01)
+    assert np.allclose(new["sig"]["d2"].independent, correct["sig"]["d2"].independent)
+
+
+if __name__ == "__main__":
+    test()
+    test_ltol_with_gtol()
